requirements.txt
absl-py==1.2.0
asgiref==3.4.1
astor==0.8.1
backcall==0.2.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2022.6.15.1
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.5
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
Django==3.2.15
entrypoints==0.4
Flask==2.0.3
gast==0.2.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.1
h5py==2.10.0
idna==3.3
image==1.5.33
imageio==2.15.0
importlib-metadata==4.8.3
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==3.0.3
jupyter-client==7.1.2
jupyter-core==4.9.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.3.4
mysql-connector-python==8.0.30
nest-asyncio==1.5.5
networkx==2.5.1
numpy==1.19.5
oauthlib==3.2.1
opencv-python==4.6.0.66
opt-einsum==3.3.0
parso==0.7.1
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.31
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.1.1
pywin32==304
pyzmq==23.2.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.17.2
scipy==1.5.4
six==1.16.0
sqlparse==0.4.2
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
tifffile==2020.9.3
tornado==6.1
traitlets==4.3.3
typing_extensions==4.1.1
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.0.3
wrapt==1.14.1
zipp==3.6.0
